Reorganize layout implementation with better iterator abstractions
